# internal/rest/openapi3.yaml
...
openapi: 3.0.0
components:
...
  schemas:
...
    Costume:
      ref: ""
      value:
        extensionprops: {}
        type: object
        properties:
          clothes:
            ref: ""
            value:
              extensionprops: {}
              type: array
              items:
                ref: ""
                value:
                  extensionprops: {}
                  type: object
                  properties:
                    id:
                      ref: ""
                      value:
                        extensionprops: {}
                        type: integer
                    name:
                      ref: ""
                      value:
                        extensionprops: {}
                        type: string
                    type:
                      ref: ""
                      value:
                        extensionprops: {}
                        type: string
          description:
            ref: ""
            value:
              extensionprops: {}
              type: string
              nullable: true
          designer:
            ref: ""
            value:
              extensionprops: {}
              type: string
          id:
            ref: ""
            value:
              extensionprops: {}
              type: integer
          images:
            ref: ""
            value:
              extensionprops: {}
              type: object
              nullable: true
              properties:
                back:
                  ref: ""
                  value:
                    extensionprops: {}
                    type: string
                    nullable: true
                details:
                  ref: ""
                  value:
                    extensionprops: {}
                    type: string
                    nullable: true
                front:
                  ref: ""
                  value:
                    extensionprops: {}
                    type: string
                    nullable: true
                sideway:
                  ref: ""
                  value:
                    extensionprops: {}
                    type: string
                    nullable: true
          isArchived:
            ref: ""
            value:
              extensionprops: {}
              type: boolean
          location:
            ref: ""
            value:
              extensionprops: {}
              type: string
              nullable: true
          name:
            ref: ""
            value:
              extensionprops: {}
              type: string
          size:
            ref: ""
            value:
              extensionprops: {}
              type: integer
              nullable: true
          tags:
            ref: ""
            value:
              extensionprops: {}
              type: object
              properties:
                colors:
                  ref: ""
                  value:
                    extensionprops: {}
                    type: array
                    items:
                      ref: ""
                      value:
                        extensionprops: {}
                        type: string
                condition:
                  ref: ""
                  value:
                    extensionprops: {}
                    type: string
                isDecor:
                  ref: ""
                  value:
                    extensionprops: {}
                    type: boolean
                materials:
                  ref: ""
                  value:
                    extensionprops: {}
                    type: array
                    items:
                      ref: ""
                      value:
                        extensionprops: {}
                        type: string
...
  responses:
...
    CreateCostumeResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Ответ при успешном создании костюма
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                type: object
                properties:
                  error:
                    ref: ""
                    value:
                      extensionprops: {}
                      nullable: true
                  result:
                    ref: ""
                    value:
                      extensionprops: {}
                      type: object
                      properties:
                        "":
                          ref: '#/components/schemas/Costume'
                          value: null
    DeleteResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Ответ при удалении чего-либо
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                type: object
                properties:
                  error:
                    ref: ""
                    value:
                      extensionprops: {}
                      nullable: true
                  result:
                    ref: ""
                    value:
                      extensionprops: {}
                      type: integer
    ErrorResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Ответ при любых ошибках
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                type: object
                properties:
                  error:
                    ref: ""
                    value:
                      extensionprops: {}
                      type: string
                  result:
                    ref: ""
                    value:
                      extensionprops: {}
                      nullable: true
...
    GetCostumesByPageResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Ответ при запросе на получение костюмов по странице
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                type: object
                properties:
                  error:
                    ref: ""
                    value:
                      extensionprops: {}
                      nullable: true
                  result:
                    ref: ""
                    value:
                      extensionprops: {}
                      type: array
                      items:
                        ref: ""
                        value:
                          extensionprops: {}
                          type: object
                          properties:
                            "":
                              ref: '#/components/schemas/Costume'
                              value: null
...
info:
  extensionprops: {}
  title: Theater API
  version: "0.1"
paths:
...
  /api/v1/costumes:
    extensionprops: {}
    post:
      extensionprops: {}
      operationId: CreateCostume
      responses:
        "200":
          ref: '#/components/responses/CreateCostumeResponse'
          value: null
        "422":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /api/v1/costumes/{costumeId}:
    extensionprops: {}
    delete:
      extensionprops: {}
      operationId: DeleteCostume
      responses:
        "200":
          ref: '#/components/responses/DeleteResponse'
          value: null
        "422":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
    put:
      extensionprops: {}
      operationId: UpdateCostume
      responses:
        "200":
          ref: '#/components/responses/CreateCostumeResponse'
          value: null
        "422":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
    parameters:
    - ref: ""
      value:
        extensionprops: {}
        name: costumeId
        in: path
        required: true
        schema:
          ref: ""
          value:
            extensionprops: {}
            type: integer
  /api/v1/costumes/{costumesPage}:
    extensionprops: {}
    get:
      extensionprops: {}
      operationId: GetCostumesByPage
      responses:
        "200":
          ref: '#/components/responses/GetCostumesByPageResponse'
          value: null
        "422":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
    parameters:
    - ref: ""
      value:
        extensionprops: {}
        name: costumesPage
        in: path
        required: true
        schema:
          ref: ""
          value:
            extensionprops: {}
            type: integer
servers:
- extensionprops: {}
  url: http://localhost:8000/
  description: Local dev